refactor(calc): extract DefenderMonHeader, simplify CalcDefenderStats to pure spread
